+++ app/Modules/Expenses/Controllers/ExpenseEditController.php
@@ -13,6 +13,7 @@ namespace BT\Modules\Expenses\Controllers;
 
 use BT\Http\Controllers\Controller;
 use BT\Modules\Categories\Models\Category;
+use BT\Modules\Clients\Models\Client;
 use BT\Modules\CompanyProfiles\Models\CompanyProfile;
 use BT\Modules\CustomFields\Models\CustomField;
 use BT\Modules\Expenses\Models\Expense;
@@ -44,6 +45,20 @@ class ExpenseEditController extends Controller
         $record['expense_date'] = DateFormatter::unformat($record['expense_date']);
         $record['amount'] = NumberFormatter::unformat($record['amount']);
         $record['tax'] = ($record['tax']) ? NumberFormatter::unformat($record['tax']) : 0;
+
+        $record['category_id'] = Category::firstOrCreate(['name' => $request->category_name])->id;
+
+        if ($request->vendor_name) {
+            $record['vendor_id'] = Vendor::firstOrCreate(['name' => $request->vendor_name])->id;
+        } else {
+            $record['vendor_id'] = 0;
+        }
+
+        if ($request->client_name) {
+            $record['client_id'] = Client::firstOrCreateByUniqueName($request->client_name)->id;
+        } else {
+            $record['client_id'] = 0;
+        }
 
         $expense = Expense::find($id);