+++ app/Modules/Documentation/Views/documentation/Release-Notes.md
@@ -5,6 +5,7 @@ Release Notes
 ### BillingTrack 5.3.0 (2021-02-08)
 - upgrade to Laravel 8 (PHP 8 compatability)
 - upgrade dependencies
+- fixed an issue with Expense Creation
 
 ### BillingTrack 5.2.0 (2020-08-01)
 - upgrade to Laravel 7